Finalizado opciones busqueda libros y mostrar libros guardados

+++ src/main/java/com/one/challenge_literalura/model/DatosLibros.java
@@ -11,7 +11,7 @@ public record DatosLibros(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") Set<DatosAutor> autor,
         @JsonAlias("languages") Set<String> idiomas,
-        @JsonAlias("download_count") Double numeroDeDescargas
+        @JsonAlias("download_count") Long numeroDeDescargas
 
 ) {
 }
